refactor: modal 컴포넌트 절대경로로 변경
