+++ src/components/modal/ModalRoot.tsx
@@ -2,8 +2,8 @@
 
 import React, { useEffect } from 'react';
 
-import { ModalProvider, useModalDispatch, useModalState } from './ModalContext';
-import ModalPortal from './ModalPortal';
+import { ModalProvider, useModalDispatch, useModalState } from '@/components/modal/ModalContext';
+import ModalPortal from '@/components/modal/ModalPortal';
 
 interface ModalRootProps {
   children: React.ReactNode;
